topologiser/requirements.txt
protobuf==3.9.1
python-bitcoinlib==0.10.2
multiaddr==0.0.8
requests=2.22.0
Flask==1.1.1

